+++ src/light_compressor/compressors/__init__.py
@@ -3,7 +3,7 @@
 from . import levels as CompressionLevel
 from .gzip import GZIPCompressor
 from .lz4 import LZ4Compressor
-from .snappy import SNAPPYCompressor
+from .snappy import SNAPCompressor
 from .zstd import ZSTDCompressor
 
 
@@ -11,6 +11,6 @@ __all__ = (
     "CompressionLevel",
     "GZIPCompressor",
     "LZ4Compressor",
-    "SNAPPYCompressor",
+    "SNAPCompressor",
     "ZSTDCompressor",
 )