+++ src/light_compressor/decompressors/__init__.py
@@ -1,9 +1,9 @@
 """Simple stream readers for compressed buffers."""
 
-from .decompress_reader import DecompressReader
+from .decompressor import DecompressReader
 from .gzip import GZIPDecompressor
 from .lz4 import LZ4Decompressor
-from .snappy import SNAPPYDecompressor
+from .snappy import SNAPDecompressor
 from .zstd import ZSTDDecompressor
 
 
@@ -11,6 +11,6 @@ __all__ = (
     "DecompressReader",
     "GZIPDecompressor",
     "LZ4Decompressor",
-    "SNAPPYDecompressor",
+    "SNAPDecompressor",
     "ZSTDDecompressor",
 )