+++ src/light_compressor/writer.py
@@ -6,7 +6,7 @@ from .compressors import (
     CompressionLevel,
     GZIPCompressor,
     LZ4Compressor,
-    SNAPPYCompressor,
+    SNAPCompressor,
     ZSTDCompressor,
 )
 
@@ -18,18 +18,15 @@ def define_writer(
 ) -> Generator[bytes, None, None]:
     """Select current method for stream object."""
 
+    if compressor_method.__class__ != CompressionMethod:
+        raise ValueError(
+            f"Unsupported compression method {compressor_method}",
+        )
+
     if compressor_method is CompressionMethod.NONE:
         return bytes_data
 
-    if compressor_method is CompressionMethod.GZIP:
-        Compressor = GZIPCompressor
-    elif compressor_method is CompressionMethod.LZ4:
-        Compressor = LZ4Compressor
-    elif compressor_method is CompressionMethod.SNAPPY:
-        Compressor = SNAPPYCompressor
-    elif compressor_method is CompressionMethod.ZSTD:
-        Compressor = ZSTDCompressor
-    else:
-        raise ValueError(f"Unsupported compression method {compressor_method}")
-
-    return Compressor(compressor_level).send_chunks(bytes_data)
+    compressor: (
+        GZIPCompressor | LZ4Compressor | SNAPCompressor | ZSTDCompressor
+    ) = compressor_method.compressor(compressor_level)
+    return compressor.send_chunks(bytes_data)